+++ NEWS.md
@@ -1,5 +1,21 @@
 # CHANGES IN bookdown VERSION 0.23
 
+- In `bs4_book`, math in footnotes is now rendered (@mine-cetinkaya-rundel, #1026)
+
+- `repo` specification in `bs4_book()` can now be done in a more flexible way: base url, branch name, subdir and icon can be specify. See `?bookdown::bs4_book()` for details (thanks, @maelle, #1036).
+
+- The `bookdown::gitbook` output format now supports an alternative search engine, namely `fuse.js`, which has several advantages over `lunr.js`, the previous search engine for `gitbook`. Using `fuse.js` will fix a number of long-standing issues such as #734, #735, and #792. To enable `fuse.js`, set the search engine to be `fuse` in `gitbook`'s config in YAML, e.g.,
+
+  ```yaml
+  output:
+    bookdown::gitbook:
+      config:
+        search:
+          engine: fuse
+  ```
+
+  Depending on user feedback, we may set `fuse` to be the default search engine in a future version of **bookdown**. We will appreciate your testing and feedback!
+
 - Figure reference links now point correctly to the top of figures (thanks, @GuillaumeBiessy, #1155).
 
 - `epub_version` argument in `epub_book()` can now be set to `epub2` to creat EPUB book of version 2. This follows an old change for default behavior in Pandoc 2.0 where the alias `epub` defaults to `epub3` and no more `epub2` (thanks, jtbayly, #1150).